refactor: Component Scan -> Java Configuration으로 Bean 등록
