+++ README.md
@@ -6,4 +6,4 @@
 ### 프로그래밍 요구사항
 #### auth 패키지를 nextstep 패키지로부터 분리
 - [x] auth 패키지 -> nextstep 패키지로 의존하는 부분을 제거한다.
-- [ ] auth 패키지에서 Spring Bean으로 사용되는 객체를 Component Scan이 아닌 `Java Configuration`으로 Bean 등록한다.
+- [x] auth 패키지에서 Spring Bean으로 사용되는 객체를 Component Scan이 아닌 `Java Configuration`으로 Bean 등록한다.
