+++ src/main/java/auth/JwtTokenProvider.java
@@ -7,9 +7,7 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 
-@Component
 public class JwtTokenProvider {
 
     @Value("${security.jwt.token.secret-key}")
