+++ src/main/java/nextstep/config/WebConfig.java
@@ -1,0 +1,24 @@
+package nextstep.config;
+
+import auth.AdminInterceptor;
+import auth.JwtTokenProvider;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Bean
+    public JwtTokenProvider jwtTokenProvider() {
+        return new JwtTokenProvider();
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry interceptorRegistry) {
+        interceptorRegistry
+            .addInterceptor(new AdminInterceptor(jwtTokenProvider()))
+            .addPathPatterns("/admin/**");
+    }
+}
