+++ src/test/java/nextstep/auth/AuthE2ETest.java
@@ -1,0 +1,53 @@
+package nextstep.auth;
+
+import auth.TokenRequest;
+import auth.TokenResponse;
+import io.restassured.RestAssured;
+import nextstep.member.MemberRequest;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.test.annotation.DirtiesContext;
+
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+class AuthE2ETest {
+
+    public static final String USERNAME = "username";
+    public static final String PASSWORD = "password";
+
+    private Long memberId;
+
+    @BeforeEach
+    void setUp() {
+        MemberRequest body = new MemberRequest(USERNAME, PASSWORD, "name", "010-1234-5678",
+            "ADMIN");
+        RestAssured
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(body)
+            .when().post("/members")
+            .then().log().all()
+            .statusCode(HttpStatus.CREATED.value());
+    }
+
+    @DisplayName("토큰을 생성한다")
+    @Test
+    void create() {
+        TokenRequest body = new TokenRequest(USERNAME, PASSWORD);
+        var response = RestAssured
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(body)
+            .when().post("/login/token")
+            .then().log().all()
+            .statusCode(HttpStatus.OK.value())
+            .extract();
+
+        Assertions.assertThat(response.as(TokenResponse.class)).isNotNull();
+    }
+}
